https://github.com/singnet/snet-daemon/issues/175 Minor changes to keep the proto simple
